requirements.txt
flwr
torch
torchvision
numpy
matplotlib
Pillow
h5py
tqdm
tensorboard
torchsummary
